Only call NetUpdate once
This doesn't need to be in a bunch of rendering code, simply only run it in TryRunTics.

Worst case scenario: if this sucks then it means dedicated servers have also sucked the whole time and we should come up with a new solution to fix both of them.

More likely scenario: nothing changes and we now aren't doing networking inside of the renderer.