Flameshield rework
The idea is to return it as a meter item while keeping and putting a restriciton on the speed building and storing part from the previous version. Using the boost while its active will drain time from the main timer to fill up store as it were, this store is what handles the boost duration which caps at 2 seconds. The speed increase from using the flame shield more and more now diminishes and caps around 81 fu/t on knuckles for example and is done seprately from the main drain timer and the store.